requirements.txt
Pillow~=9.0.1
quart~=0.16.3
gunicorn~=20.1.0
uvicorn[standard]~=0.17.4
aiohttp~=^3.8.1